Finish provisioning brains via ansible.
Two roles created:
 * A base Node RED install
 * A Node RED install with an RFID reader
Then sets up the MiaB brain on top of that.
Boot progress script moved to systemd (in line with newer Raspbian)
NDEF reading C program from simple_sl030_rfid_reader compiled and included in the miab/ folder, along with a service to copy the correct one to a common filename on boot (so will work regardless of the type of Pi the SD card is placed into).
FIXME However, it still needs the Pi 1 and Pi 2 versions compiling and adding.